FloatState / H2O / H2O2 / MessageBoxDialog: Fix 64 bit build warnings

+++ Src/Common/FloatState.cpp
@@ -270,7 +270,7 @@ LRESULT CFloatState::CallWindowProc(WNDPROC pfnSuper, HWND hWnd, UINT uMsg, WPAR
 	switch (uMsg)
 	{
 	case WM_NCHITTEST:
-		lResult = CFloatState::AdjustHit(lResult);
+		lResult = CFloatState::AdjustHit(static_cast<UINT>(lResult));
 		break;
 	case WM_WINDOWPOSCHANGING:
 		CFloatState::Float(reinterpret_cast<WINDOWPOS *>(lParam));
